+++ libraries/ui/src/main/java/androidx/media3/ui/PlayerView.java
@@ -333,6 +333,7 @@ public class PlayerView extends FrameLayout implements AdViewProvider {
   private boolean controllerAutoShow;
   private boolean controllerHideDuringAds;
   private boolean controllerHideOnTouch;
+  private boolean enableComposeSurfaceSyncWorkaround;
 
   public PlayerView(Context context) {
     this(context, /* attrs= */ null);
@@ -1305,6 +1306,19 @@ public class PlayerView extends FrameLayout implements AdViewProvider {
   }
 
   /**
+   * Whether to enable a workaround for the Compose {@code AndroidView} and {@link SurfaceView}
+   * compatibility issue described in <a
+   * href="https://github.com/androidx/media/issues/1237">androidx/media#1237</a>.
+   *
+   * <p>This workaround causes issues with shared element transitions in XML views, so is disabled
+   * by default (<a href="https://github.com/androidx/media/issues/1594">androidx/media#1594</a>).
+   */
+  @UnstableApi
+  public void setEnableComposeSurfaceSyncWorkaround(boolean enableComposeSurfaceSyncWorkaround) {
+    this.enableComposeSurfaceSyncWorkaround = enableComposeSurfaceSyncWorkaround;
+  }
+
+  /**
    * Gets the view onto which video is rendered. This is a:
    *
    * <ul>
@@ -1758,7 +1772,7 @@ public class PlayerView extends FrameLayout implements AdViewProvider {
   @Override
   protected void dispatchDraw(Canvas canvas) {
     super.dispatchDraw(canvas);
-    if (Util.SDK_INT == 34 && surfaceSyncGroupV34 != null) {
+    if (Util.SDK_INT == 34 && surfaceSyncGroupV34 != null && enableComposeSurfaceSyncWorkaround) {
       surfaceSyncGroupV34.maybeMarkSyncReadyAndClear();
     }
   }
@@ -1830,7 +1844,9 @@ public class PlayerView extends FrameLayout implements AdViewProvider {
 
     @Override
     public void onSurfaceSizeChanged(int width, int height) {
-      if (Util.SDK_INT == 34 && surfaceView instanceof SurfaceView) {
+      if (Util.SDK_INT == 34
+          && surfaceView instanceof SurfaceView
+          && enableComposeSurfaceSyncWorkaround) {
         // Register a SurfaceSyncGroup to work around https://github.com/androidx/media/issues/1237
         // (only present on API 34, fixed on API 35).
         checkNotNull(surfaceSyncGroupV34)